Move `createHTTPAction` action to `writeback`
