+++ frontend/src/metabase/writeback/actions.ts
@@ -1,5 +1,24 @@
+import { t } from "ttag";
+import { push } from "react-router-redux";
+
+import Actions from "metabase/entities/actions";
 import { ActionsApi } from "metabase/services";
+import { addUndo } from "metabase/redux/undo";
+
 import Table from "metabase-lib/lib/metadata/Table";
+
+import {
+  Parameter,
+  ParameterId,
+  ParameterTarget,
+} from "metabase-types/types/Parameter";
+import { Dispatch, GetState } from "metabase-types/store";
+
+import {
+  HttpActionErrorHandle,
+  HttpActionResponseHandle,
+  HttpActionTemplate,
+} from "./types";
 
 export type InsertRowPayload = {
   table: Table;
@@ -97,3 +116,61 @@ export const deleteManyRows = (payload: BulkDeletePayload) => {
     { bodyParamName: "body" },
   );
 };
+
+export type CreateHttpActionPayload = {
+  name: string;
+  description: string;
+  template: HttpActionTemplate;
+  response_handle: HttpActionResponseHandle;
+  error_handle: HttpActionErrorHandle;
+  parameters: Record<ParameterId, Parameter>;
+  parameter_mappings: Record<ParameterId, ParameterTarget>;
+};
+
+export const createHttpAction =
+  (payload: CreateHttpActionPayload) =>
+  async (dispatch: Dispatch, getState: GetState) => {
+    const {
+      name,
+      description,
+      template,
+      error_handle = null,
+      response_handle = null,
+      parameters,
+      parameter_mappings,
+    } = payload;
+    const data = {
+      method: template.method || "GET",
+      url: template.url,
+      body: template.body || JSON.stringify({}),
+      headers: JSON.stringify(template.headers || {}),
+      parameters: template.parameters || {},
+      parameter_mappings: template.parameter_mappings || {},
+    };
+    const newAction = {
+      name,
+      type: "http",
+      description,
+      template: data,
+      error_handle,
+      response_handle,
+      parameters,
+      parameter_mappings,
+    };
+    const response = await dispatch(Actions.actions.create(newAction));
+    const action = Actions.HACK_getObjectFromAction(response);
+    if (action.id) {
+      dispatch(
+        addUndo({
+          message: t`Action saved!`,
+        }),
+      );
+      dispatch(push(`/action/${action.id}`));
+    } else {
+      dispatch(
+        addUndo({
+          message: t`Could not save action`,
+        }),
+      );
+    }
+  };
